Make a separate package for the runner
